+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## v1.2.0 2026-05-30
 
 Added:
+- added ptr access to struct fields and passing by reference with &
 - added varargs ... 
 - expanded io stdlib